requirements.txt
fastapi
uvicorn
openai
httpx
python-dotenv
pydantic
langchain
twilio
tenacity
